Reference

References

msJ.builoto

This page lists all the documented elements of the msJ.jl package covering all modules and submodules.

Contents

References

Contents

Main module

Types

I/O

Filtering

Extracting subsets

Process

Plots

Utilities

Main module

msJ.msJ — Module.

Main module for msJ.jl– A Julia package to load and process mass spectrometry data.

source

Types

Submodule with types and structures used to stored the data and dispatch to the right methods.

Data types

msJ.MScontainer — Type.

abstract type MScontainer  end

Abstract type containing any imported data belongs to the MScontainer type.

source

msJ.MSscan — Type.

struct MSscan <: MScontainer

Data structure used to store individual mass spectrometry scans organized following the structure of mzXML files.

struct MSscan <: MScontainer num::Int # num rt::Float64 # retentionTime tic::Float64 # totIonCurrent mz::Vector{Float64} # m/z int::Vector{Float64} # intensity level::Int # msLevel basePeakMz::Float64 # basePeakMz basePeakIntensity::Float64 # basePeakIntensity precursor::Float64 # precursorMz polarity::String # polarity activationMethod::String # activationMethod collisionEnergy::Float64 # collisionEnergy end

source

msJ.MSscans — Type.

struct MSscans  <: MScontainer

Data structure designed to store mass spectra obtained after filtering operation along with the history of these operation.

struct MSscans <: MScontainer num::Vector{Int} # num rt::Vector{Float64} # retentionTime tic::Float64 # totIonCurrent mz::Vector{Float64} # m/z int::Vector{Float64} # intensity level::Vector{Int} # msLevel basePeakMz::Float64 # basePeakMz basePeakIntensity::Float64 # basePeakIntensity precursor::Vector{Float64} # precursorMz polarity::Vector{String} # polarity activationMethod::Vector{String} # activationMethod collisionEnergy::Vector{Float64} # collisionEnergy s::Vector{Float64} # variance end

source

msJ.Chromatogram — Type.

struct Chromatogram  <: MScontainer

Data structure used to retrieve chromatography data.

struct Chromatogram <: MScontainer rt::Vector{Float64} ic::Vector{Float64} maxic::Float64 end

source

Methods Types

msJ.MethodType — Type.

abstract type MethodType  end

Type containing all the methods used for filtering the data.

source

msJ.BasePeak — Type.

struct BasePeak <: MethodType

Structure for multiple dispatching to retrieve base peak chromatogram.

source

msJ.TIC — Type.

struct TIC <: MethodType

Dispatching to retrieve total ion current chromatogram.

source

msJ.∆MZ — Type.

struct ∆MZ{argT <: Union{Real, AbstractVector{<:Real} }} <: MethodType

Structure for multiple dispatching to retrieve extracted ion current chromatogram around an m/z ± ∆mz value given by arg = [mz, ∆mz]

source

msJ.MZ — Type.

struct MZ{argT <: Union{Real, AbstractVector{<:Real} }} <: MethodType

Structure for multiple dispatching to retrieve extracted ion current chromatogram around for m/z in the range arg = [mz1, mz2].

source

msJ.SG — Type.

struct SG{argT <: Int} <: MethodType   #Savinsky & Golay filtering

Structure for multiple dispatching to Savinsky & Golay filtering, providing the order, window size and derivative to be performed. Defaults values are provided in functions calls.

source

msJ.TBPD — Type.

struct TBPD{argT1 <: Symbol, argT2 <: Real}  <: MethodType

Structure for multiple dispatching to Template Base Peak Detection centroiding, providing the shape of the template function, the resolution and threshold. Defaults values are provided in functions calls.

source

msJ.SNRA — Type.

struct SNRA{argT1 <: Real, argT2 <: Int}  <: MethodType

Structure for multiple dispatching to Signal to Noise Ratio Analysis centroiding, providing the threshold value and the size of the region. Defaults values are provided in functions calls.

source

msJ.TopHat — Type.

TopHat{argT <: Int} <: MethodType

Structure for multiple dispatching to TopHat baseline correction. Region is used specify the dimention over which this operation performed

source

msJ.LOESS — Type.

LOESS{argT <: Int} <: MethodType

Structure for multiple dispatching to LOcally Weighted Error Sum of Squares regression (LOESS) baseline correction.

source

msJ.IPSA — Type.

struct IPSA{argT1 <: Int, argT2 <: Real} <: MethodType

Structure for multiple dispatching to iterative polynomial smoothing algorithm (IPSA) baseline correction.

source

Filters

msJ.FilterType — Type.

abstract type FilterType end

This type contains the structures for filtering the data.

source

msJ.RT — Type.

RT{argT <: Union{Real, AbstractVector{<:Real},  AbstractVector{<:AbstractVector{<:Real}} }}

This type contains the structures for filtering the data.

source

msJ.IC — Type.

struct IC{argT <: Union{Real, AbstractVector{<:Real} }} <: FilterType

Used for multiple dispatching to Template Base Peak Detection centroiding, providing the shape of the template function, the resolution and threshold. Defaults values are provided in functions calls.

source

msJ.Level — Type.

struct Level{argT <: Union{Int, AbstractVector{<:Int} }} <: FilterType

Used to dispatch filters to MS level.

source

msJ.Scan — Type.

 Scan{argT <: Union{Int, AbstractVector{<:Int} }} <: FilterType

Dispatch filter to scan num.

source

msJ.Polarity — Type.

struct Polarity{argT <: Union{String, AbstractVector{<:String} }} <: FilterType

Dispatch filter to polarity.

source

msJ.Activation_Method — Type.

struct Activation_Method{argT <: Union{String, AbstractVector{<:String} }} <: FilterType

Dispatch filter to activation methods

source

msJ.Activation_Energy — Type.

struct Activation_Energy{argT <: Union{Real, AbstractVector{<:Real} }} <: FilterType

Dispatch filter to activation energies.

source

msJ.Precursor — Type.

struct Precursor{argT <: Union{Real, AbstractVector{<:Real} }} <: FilterType

Dispatch filter to precursor.

source

I/O

Module for importing and exporting data. Dispatch to specific methods according to the file extension

msJ.info — Method.

info(filename::String, verbose::Bool = false)

The function looks into an file and returns in an Array{String} containing the number of scans and the different scans described by their MS level, polarity and eventually the precursor m/z followed by the activation method and collision energy. Each entry is unique, which gives a summary of the input file. With verbose = true, the functions also returns the parentFile, msManufacturer, msModel, msIonisation, msMassAnalyzer, msDetector, software and dataProcessing if existing.

Example

julia> info("test1.mzXML")
4-element Array{String,1}:
 "51 scans"               
 "MS1+"                   
 "MS2+ 1255.5  CID(CE=18)"
 "MS3+ 902.33  PQD(CE=35)"
julia> msJ.info("test1.mzXML", verbose = true)
12-element Array{String,1}:
 "parentFile: test1_msJ_1.raw"
 "msManufacturer: Thermo Finnigan"
 "msModel: LTQ XL"      
 "msIonisation: ESI"
 "msMassAnalyzer: ITMS"
 "msDetector: unknown"
 "software: Xcalibur, 2.6.0 SP3"
 "dataProcessing: conversion, ReAdW 4.3.1(build Sep  9 2009 12:30:29)"
 "51 scans"
 "MS1+"
 "MS2+ 1255.5  CID(CE=18)"
 "MS3+ 902.33  PQD(CE=35)"

source

msJ.load — Method.

function load(filename::String)

Checks the file extension and calls the right function to load the mass spectra if it exists. Returns an array of msJ.MSscan where the individual mass spectra are stored.

Examples

julia> scans = load("test.mzXML")
51-element Array{msJ.MSscan,1}:
 msJ.MSscan(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ....

source

msJ.retention_time — Method.

retention_time(filename::String)

Returns an array composed of the retention times of the individual mass spectra.

Examples

julia> retention_time("test.mzXML")
51-element Array{Float64,1}:
  0.1384
  0.7307
  2.1379
....

source

msJ.chromatogram — Method.

chromatogram(filename::String, filters::FilterType...; method::MethodType=TIC())

Returns a structure holding the retention time (rt), the ion current (ic) and the maximum value (maxic) for all the mass spectra within the file. Alternatively, other options may be supplied such as method = msJ.BasePeak, which returns the base peak intensity, method = msJ.∆MZ([500,5]), which returns the ion current for the range mz = 500 ± 5, or method = msJ.MZ([200,1000]) which return the ion current in the range from m/z 200 to m/z 1000. The data may be filtered by ms level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method")...

Examples

julia> rt, ic = chromatogram("test.mzxml")
([0.1384  …  60.4793], [4.74795e6  …  17.4918])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.BasePeak() )
([0.1384  …  60.4793], [102558.0  …  1.23181])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.∆MZ([500,5]) )
([0.1384  …  60.4793], [46036.6  …  14.2529])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.MZ([200,1000]))
([0.1384  …  60.4793], [4.74795e6  …  17.4918])

source

msJ.average — Method.

average(filename::String, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within file. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

mzXML

Interface to the mzxml file format

msJ.info_mzxml — Function.

info_mzxml(filename::String, info::Vector{String}, verbose::Bool=false)

Returns the information content of an mzXML file into a string. Verbosity is controlled by the verbose Boolean variable set by default to false.

source

msJ.load_mzxml_all — Function.

load_mzxml_all(filename::String)

Load an entire an mzxml file, and return the

source

msJ.load_mzxml — Function.

load_mzxml(filename::String, index::Int

Load from an mzxml file the scan num that match the input index

source

msJ.load_mzxml_spectrum — Function.

load_mzxml_spectrum(c::XMLElement)

From an XMLElement, returns the data into an MSscan.

source

msJ.retention_time — Method.

retention_time(msRun::XMLElement)

From an XMLE element returns the retention time.

source

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Level{<:Int}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Level{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Scan{<:Int}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Scan{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Polarity{<:String}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Polarity{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::RT{<:Real}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::RT{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::RT{<:AbstractVector{<:AbstractVector} } ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::IC{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Precursor{<:Real}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Precursor{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Activation_Energy{<:Real}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Activation_Energy{<:AbstractVector}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Activation_Method{<:String}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for msJ.average(msRun::XMLElement, argument::Activation_Method{<:AbstractVector}). Check Documenter's build log for details.

msJ.extracted_chromatogram — Method.

extracted_chromatogram(filename::String, indices::Vector{Int},method::MethodType)

Returns the extracted chromatogram for input file according to the selected method and for set of scan num as input

source

msJ.composite_spectra — Method.

composite_spectra(filename::String, indices::Vector{Int}, stats::Bool)

Returns the average MSscans for input filename and according to the input scan num. Calculation of variance is controlled by the stats Boolean variable.

source

Filtering

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.chromatogram — Method.

chromatogram(scans::Vector{MSscan}, filters::FilterType...; method::MethodType=TIC())

Returns the retention time and the total ion current by default for all the mass spectra within the Array of mass spectrum container MSscan. Alternatively, other options may be supplied such as method = msJ.BasePeak, which returs the base peak intensity, method = msJ.∆MZ([500,5]), which returns the ion current for the range mz = 500 ± 5, or method = msJ.MZ([200,1000]) which return the ion current in the range from m/z 200 to m/z 1000. The data may be filtered by ms level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method")...

Examples

julia> rt, ic = chromatogram("test.mzxml")
([0.1384  …  60.4793], [4.74795e6  …  17.4918])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.BasePeak() )
([0.1384  …  60.4793], [102558.0  …  1.23181])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.∆MZ([500,5]) )
([0.1384  …  60.4793], [46036.6  …  14.2529])
julia> rt, ic = chromatogram("test.mzxml", method = msJ.MZ([200,1000]))
([0.1384  …  60.4793], [4.74795e6  …  17.4918])

source

msJ.retention_time — Method.

retention_time(scans::Vector{MSscan})

Returns an array composed of the retention times of the individual mass spectra.

Examples

julia> retention_time("scans")
51-element Array{Float64,1}:
  0.1384
  0.7307
  2.1379
....

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.average — Method.

average(scans::Vector{MSscan}, arguments::FilterType...; stats::Bool=true)

Returns the average mass spectrum container (MSscans) along with the sample standard deviation of the intensities with stats=true (default) for all the mass spectra within the Array of mass spectrum container MSscan.. The data may be filtered by level, precursor mass, activation methods, etc, using the arguments msJ.Level(N), msJ.Precursor(mz), msJ.Activation_Method("method"), or any combination of these arguments.

Examples

julia> spectrum = average("test.mzxml")
msJ.MSscans([1, 2, 3 ....
julia> spectrum = average("test.mzxml", msJ.Level(1) )
msJ.MSscans([1, 4, 7, 10,
julia> spectrum = average("test.mzxml", msJ.Precursor(1255.5) )
msJ.MSscans([2, 5, 8, 11, ...
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD") )
msJ.MSscans([3, 6, 9, 12, 15,
julia> spectrum = average("test.mzxml", msJ.Activation_Method("PQD"), msJ.Polarity("+"), msJ.RT([10,20]))
msJ.MSscans([9, 12, 15, 18], ...

source

msJ.extracted_chromatogram — Method.

extracted_chromatogram(scans::Vector{MSscan}, indices::Vector{Int},method::MethodType)

Returns the extracted chromatogram for input Array of MSscan according to the selected method and for set of scan num as input

source

msJ.composite_spectra — Method.

composite_spectra(scans::Vector{MSscan}, indices::Vector{Int}, stats::Bool)

Returns the average MSscans for input Array of MSscan and according to the input scan num. Calculation of variance is controlled by the stats Boolean variable.

source

Extracting subsets

msJ.extract — Method.

extract(filename::String, arguments::FilterType...)

Search for scans matching the argument MS level and returns an array of matching MSscans otherwise returns an ErrorException: "No matching spectra found."

Examples

julia> sub_set = extract("test.mzxml")
6-element Array{msJ.MSscan,1}:
 msJ.MSscan(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ....
julia> sub_set = extract("test.mzxml", msJ.Level(2) )      # extract MS/MS spectra
msJ.MSscan(2, 0.7307, 9727.2, [345.083, 345.167, 345.25, 345.333, 345.417, 345.5, 345.583, 345.667, 345.75, 345.833  …  1999.25, 1999.33, 1999.42, 1999.5, 1999.58 ....
 msJ.MSscan(5, 4.3442, 12203.5, [345.083, 345.167, 345.25, 345.333, 345.417, 345.5, 345.583, 345.667, 345.75, 345.833  …  1999.25, 1999.33, 1999.42, 1999.5, 1999.58, ....

source

msJ.extract — Method.

extract(scans::Vector{MSscan}, arguments::FilterType...)

Search for scans matching the argument MS level and returns an array of matching MSscans otherwise returns an ErrorException: "No matching spectra found."

Examples

julia> scans = load("test.mzxml")                          # load mass spectra
6-element Array{msJ.MSscan,1}:
 msJ.MSscan(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ...
julia> sub_set = extract(scans)                            # extract a sub_set without conditions returns the original data
6-element Array{msJ.MSscan,1}:
 msJ.MSscan(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ....
julia> sub_set = extract(scans, msJ.Level(2) )      # extract MS/MS spectra
msJ.MSscan(2, 0.7307, 9727.2, [345.083, 345.167, 345.25, 345.333, 345.417, 345.5, 345.583, 345.667, 345.75, 345.833  …  1999.25, 1999.33, 1999.42, 1999.5, 1999.58 ....
 msJ.MSscan(5, 4.3442, 12203.5, [345.083, 345.167, 345.25, 345.333, 345.417, 345.5, 345.583, 345.667, 345.75, 345.833  …  1999.25, 1999.33, 1999.42, 1999.5, 1999.58, ....

source

msJ.build_subset — Method.

build_subset(filename::String, indices::Vector{Int})

Returns a Vector of MSscan from the input file according to the scan num (indices).

source

msJ.build_subset — Method.

build_subset(scans::Vector{MSscan}, indices::Vector{Int})

Returns a Vector of MSscan according to the input scan num.

source

Process

Mass spectrum

msJ.smooth — Method.

smooth(scan::MScontainer; method::MethodType=SG(5, 9))

Smooth the intensity of the input data and returns a similar structure.

Examples

julia> smoothed_data = msJ.smooth(scans)
msJ.MSscans(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ....

source

msJ.smooth — Method.

smooth(scans::Vector{MSscan}; method::MethodType=SG(5, 9, 0))

Smooth the intensity of the input data and returns a similar structure.

Examples

julia> scans = load("filename")
julia> smoothed_data = msJ.smooth(scans)
6-element Array{msJ.MSscan,1}:
 msJ.MSscan(1, 0.1384, 5.08195e6 .....

source

msJ.savitzky_golay_filtering — Method.

savitzky_golay_filtering(scan::msJ.MScontainer, order::Int, window::Int, deriv::Int)

Savinsky and Golay filtering of mz and int data within the MSscan(s) container.

source

Missing docstring.

Missing docstring for msJ.smooth(scans::Vector{MSscan}; method::MethodType=SG(5, 9, 0)). Check Documenter's build log for details.

msJ.centroid — Method.

centroid(scan::MScontainer; method::MethodType=MethodType=SNRA(1., 100) )

Peak picking algorithm taking a MSscan or MSscans object as input and returning an object of the same type containing the detected peaks. Available algorithm are : Signal to Noise Ratio (SNR) and Template Based Peak Detection (TBPD). Default method is Signal to Noise Ratio Analysis (SNRA), with default threshold = 1.0 and region = 100.

Examples

julia> centroid(scans)
MSscans(1, 0.1384, 5.08195e6, [140.083, 140.167, 140.25, 140.333, 140.417, 140.5, 140.583, 140.667, 140.75, 140.833  …  1999.25, 1999.33, 1999.42, ....

source

Chromatogram

No functions yet. To be added.

Plots

msJ.plots — Module.

Plotting module for MScontainer data type (MSscan, MSscans and Chromatogram).

julia> plot(scans[1])
julia> plot(chr)

source

RecipesBase.apply_recipe — Method.

h(cr::msJ.Chromatogram; method = :relative)

Allows plotting directly chromatograms. The defaults relative intensity plotting may be changed by setting method = :absolute.

source

RecipesBase.apply_recipe — Method.

g(ms::msJ.MSscans; method = :relative)

Allows plotting directly mass spectra MSscans. The defaults relative intensity plotting may be changed by setting method = :absolute.

source

RecipesBase.apply_recipe — Method.

f(ms::MSscan; method = :relative)

Allows plotting directly mass spectra MSscan. The defaults relative intensity plotting may be changed by setting method = :absolute.

source

msJ.plots.normalisation — Method.

normalisation(cr::msJ.Chromatogram)

Normalization function for plotting chromatograms in raltive intensity.

source

msJ.plots.normalisation — Method.

normalisation(ms::msJ.MScontainer)

Normalization function for plotting mass spectra in relative intensity.

source

msJ.plots.scaling — Method.

scaling(cr::msJ.Chromatogram)

Scaling function to display retention times of chromatograms in minutes instead of seconds.

source

Utilities

Base overloaded

Base.:+ — Method.

+(a::MScontainer, b::MScontainer)

Addition of mass spectra elementwise.

julia> scans[1] - scans[2]
msJ.MSscans([1, 2], [0.1384, 0.7307]

source

Base.:- — Method.

-(a::MScontainer, b::MScontainer)

Substraction of mass spectra elementwise. Negative scan num refers the 'b' MScontainer

julia> a - b
msJ.MSscans([1, 4], [0.1384, 3.7578, -0.1384, -3.7578]...

source

Base.:/ — Method.

/(a::MSscan, N::Real)

Divide the intensity and the tic data of a MSscan by a number

julia> scans[1] / 1.0e2
msJ.MSscan(1, 0.1384, 50819.5, [140. ....

source

Base.:/ — Method.

/(a::MSscans, N::Real)

Divide in the intenisty, tic and variance of a MSscans by a number

julia> a / 1.0e2
msJ.MSscans(1, 0.1384, 50819.5, [140. ....

source

Base.:* — Method.

*(a::MSscan, N::Real)

Multiply the intensity and the tic data of a MSscan by a number

julia> scans[1] * 1.0e2
msJ.MSscan(1, 0.1384, 50819.5, [140. ....

source

Base.:* — Method.

*(a::MSscans, N::Real)

Multiply in the intenisty, tic and variance of a MSscans by a number

julia> a * 1.0e2
msJ.MSscans(1, 0.1384, 50819.5, [140. ....

source

Base.:* — Method.

*(N::Real, a::MScontainer)

Commutation of multiplication of number with MSscontainer

source

Base.:* — Method.

*(a::MScontainer, b::MScontainer)

Multiplication of mass spectra elementwise.

julia> a * b
msJ.MSscans([2, 5], [0.7307, 4.344

source

Utility function

msJ.avg — Method.

avg(a::MScontainer, b::MScontainer)

Returns the average of the input mass spectra and compute the variance using an incremental Welford algorithm.

julia> msJ.avg(scans[1], scans[4])
msJ.MSscans([1, 4], [0.1384, 3.7578], ....

source

msJ.add_ion_current — Method.

add_ion_current(x::AbstractArray, y::AbstractArray, a::Real, b::Real)

Returns sum the ion current (int) within the m/z range defined by the a and b input values.

source

msJ.num2pnt — Method.

num2pnt(x::AbstractArray, val::Real)

General purpose utility function used to retrieve the index of an array for which the value is closest to the input.

source

msJ.savitzky_golay — Method.

savitzky_golay(int::Vector{Float64}, order::Int, window::Int, deriv::Int)

Savinsky and Golay filter removes high frequency noise from data. Parameters: int::AbstractArray order::Int order of the polynomial window::Int length of the window, has to be an odd number deriv::Int the order of the derivative to be computed. Default = 0 leads to smoothing only.

source

msJ.extremefilt — Method.

extremefilt(input::AbstractArray, minmax::Function, region::Int)

Return the erosion or the dilation of the input over the region, which the size of the structuring element.

source

msJ.morpholaplace — Method.

morpholaplace(input::AbstractArray, region::Int)

Perfoms morphological Laplacian of the input array, as defined by the addition of the dilatation and the erosion of the input array.

source

msJ.morphogradient — Method.

morphogradient(input::AbstractArray, region::Int)

Perfoms morphological Gradient of the input array, defined by the difference between the dilatation and the erosion of the input array.

source

msJ.tophat — Method.

tophat(input::AbstractArray, region::Int)

Performs the Top Hat of the input Array, defined by the difference between the input and its morphological opening.

source

msJ.bottomhat — Method.

bottomhat(input::AbstractArray, region::Int)

Performs the Bottom Hat of the input Array, defined by the difference between the morphological closing of the input and the input.

source

msJ.opening — Method.

opening(input::AbstractArray, region::Int)

Performs the morphological opening of the input Array, which is the dilatation of the erosion of the input

source

msJ.closing — Method.

closing(input::AbstractArray, region::Int

Performs the morphological closing of the input Array, which is defined as the erosion of the dilatation of the input.

source

msJ.erosion — Method.

erosion(input::AbstractArray, region::Int)

Performs the morphological erosion of the input, which is the minimum-filtering over the structuring element region.

source

msJ.dilatation — Method.

dilatation(input::AbstractArray, region::Int)

Performs the morphological dilatation of the input, which is the maximum-filtering over the structuring element region

source